enerator for bgcdsl
