+++ bgc-dsl/bgcLSPServer/fileWriter/jinja-templates/fun_group.txt
@@ -1,6 +1,7 @@
 TYPE, EXTENDS(fungroup), PUBLIC :: {{class_name}}
     {%include parameter_list.txt%}
 
+
 CONTAINS
     PROCEDURE :: flux => {{class_name}}_flux
     PROCEDURE :: read => {{class_name}}_read
